[feat] : 스니펫 페이지 구현
